+++ load-test/gulpfile.js
@@ -4,4 +4,4 @@ var builder = require('@jenkins-cd/js-builder');
 // Bundle the modules.
 // See https://github.com/jenkinsci/js-builder
 //
-builder.bundle('src/main/js/sse-gateway-sample.js').less('src/main/js/sse-gateway-sample.less');
+builder.bundle('src/main/js/sse-gateway-load.js').less('src/main/js/sse-gateway-load.less');
